Factor common code out of the bitrig and openbsd modules
The new directory structure better reflects the historical relationship
of the BSDs, and includes 400 fewer duplicative lines of code.
